requirements.txt
pandas
snowflake-connector-python
dbt-core
dbt-snowflake
faker
python-dotenv